split into different assemblies
