+++ aspnet/src/SDP.TaskManagement.WebHost/SDP.TaskManagement.WebHost.csproj
@@ -12,4 +12,8 @@
 		<PackageReference Include="Swashbuckle.AspNetCore" Version="8.1.1" />
 	</ItemGroup>
 
+	<ItemGroup>
+	  <ProjectReference Include="..\..\SDP.TaskManagement.Application\SDP.TaskManagement.Application.csproj" />
+	</ItemGroup>
+
 </Project>
